% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGE.CI.R
\name{DGE.CI}
\alias{DGE.CI}
\title{plot and get summary}
\usage{
DGE.CI(DE, FoldChange = 0, cutoff = 0.05, type.sig = "p")
}
\arguments{
\item{DE}{list of Differentially Expressed Genes from Various methods.}

\item{FoldChange}{is cuoff to mark genes as Differentially expressed}

\item{cutoff}{is either pvalue of FDR cutoff for filtering}

\item{type.sig}{vector \code{c('p', 'FDR')} default \code{'p'}}
}
\value{
data frame of Deferentially expressed genes counts in various methods
}
\description{
plot and get summary
}
\examples{

data("DEG")
DE.list<-list("edger" =dge_edger, "edgerql" = dge_edgerql,
"voom" = dge_voom )
DGE.CI(DE=DE.list , FoldChange=1.2 , cutoff =0.01 , type.sig ="FDR")


}
