% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodEdgeR.R
\name{MethodEdgeR}
\alias{MethodEdgeR}
\title{EdgeR method}
\usage{
MethodEdgeR(count.table, design.matrix)
}
\arguments{
\item{count.table}{count table}

\item{design.matrix}{design matrix}
}
\value{
data.frame of differntial gene expression quantification by EdgeR
}
\description{
EdgeR method
}
\examples{

data("count.table")
ss <- data.frame("trt" = colnames(count.table))
rownames(ss) <- ss$trt
ss$trt <- as.integer(grepl("T",ss[,1]))

design <- model.matrix(~ss$trt)


EdgeR <- MethodEdgeR(count.table = count.table, design.matrix = design)
}
