% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodEdgeRLRT.R
\name{MethodEdgeRLRT}
\alias{MethodEdgeRLRT}
\title{EdgeR LRT method}
\usage{
MethodEdgeRLRT(count.table, design.matrix)
}
\arguments{
\item{count.table}{count table}

\item{design.matrix}{design matrix}
}
\value{
data.frame of differential gene expression quantification by EdgeR
}
\description{
EdgeR LRT method
}
\examples{

data("count.table")
ss <- data.frame("trt" = colnames(count.table))
rownames(ss) <- ss$trt
ss$trt <- as.integer(grepl("T",ss[,1]))

design <- model.matrix(~ss$trt)

EdgeR.LRT <- MethodEdgeRLRT(count.table = count.table, design.matrix = design)
}
