% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodVoom.R
\name{MethodVoom}
\alias{MethodVoom}
\title{limma voom method}
\usage{
MethodVoom(count.table, design.matrix)
}
\arguments{
\item{count.table}{count table}

\item{design.matrix}{design matrix}
}
\value{
data.frame of differential gene expression quantification by Voom
}
\description{
limma voom method
}
\examples{

data("count.table")
ss <- data.frame("trt" = colnames(count.table))
rownames(ss) <- ss$trt
ss$trt <- as.integer(grepl("T",ss[,1]))
design <- model.matrix(~ss$trt)

VoomDGE <- MethodVoom(count.table = count.table , design.matrix = design)
}
