% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiVolcano.R
\name{multiVolcano}
\alias{multiVolcano}
\title{plot volcano for easy comparision}
\usage{
multiVolcano(
  DE,
  FoldChange = 0,
  cutoff = 0.05,
  DE.Only = TRUE,
  type.sig = "p",
  show.genes = FALSE,
  genes = NULL
)
}
\arguments{
\item{DE}{list of Differentially Expressed Genes from Various methods.}

\item{FoldChange}{is cuoff to mark genes as Differentially expressed}

\item{cutoff}{is either pvalue of FDR cutoff for filtering}

\item{DE.Only}{Logical, show unchanged genes or not default TRUE}

\item{type.sig}{vector \code{c('p', 'FDR')} default \code{'p'}}

\item{show.genes}{logical weather to show genes as label}

\item{genes}{vector tor of genes for which only volcano plot will be drawn}
}
\description{
plot volcano for easy comparision
}
\examples{

data("DEG")
DE.list<-list("edger" =dge_edger, "edgerql" = dge_edgerql,
                 "deseq2" = dge_deseq2, "voom" = dge_voom )
genes.list<-c("Gene7673", "Gene28034", "Gene38639")
multiVolcano(DE.list, FoldChange = 1.4, DE.Only = FALSE )
multiVolcano(DE.list, FoldChange = 1.4, DE.Only = FALSE , genes = genes.list)


}
