% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabulateDEStats.R
\name{TabulateStats}
\alias{TabulateStats}
\title{plot Prapotional venn diagram for upregulated and downregulated genes}
\usage{
TabulateStats(
  DE,
  FoldChange = 0,
  cutoff = 0.05,
  DE.Only = TRUE,
  type.sig = "p"
)
}
\arguments{
\item{DE}{list of Differentially Expressed Genes from Various methods.}

\item{FoldChange}{is cuoff to mark genes as Differentially expressed}

\item{cutoff}{is either pvalue of FDR cutoff for filtering}

\item{DE.Only}{Logical, show unchanged genes or not default TRUE}

\item{type.sig}{vector \code{c('p', 'FDR')} default \code{'p'}}
}
\value{
data frame of Deferentially expressed genes counts in various methods
}
\description{
plot Prapotional venn diagram for upregulated and downregulated genes
}
\examples{

data("DEG")
DE.list<-list("edger" =dge_edger, "edgerql" = dge_edgerql,
"deseq2" = dge_deseq2, "voom" = dge_voom )
TabulateStats(DE=DE.list , FoldChange=1.5 , cutoff =0.01 , type.sig ="FDR")


}
