% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpSetPlot.R
\name{UpSetPlot}
\alias{UpSetPlot}
\title{plot UpSetPlot diagram for upregulated and downregulated genes}
\usage{
UpSetPlot(DE, FoldChange = 0, cutoff = 0.05, type.sig = NULL)
}
\arguments{
\item{DE}{list of Differentially Expressed Genes from Various methods.}

\item{FoldChange}{is cuoff to mark genes as Differentially expressed}

\item{cutoff}{is either pvalue of FDR cutoff for filtering}

\item{type.sig}{vector \code{c('p', 'FDR')} default \code{'p'}}
}
\value{
\code{NULL}-
}
\description{
plot UpSetPlot diagram for upregulated and downregulated genes
}
\examples{

data("DEG")
DE.list<-list("edger" =dge_edger, "edgerql" = dge_edgerql,
"deseq2" = dge_deseq2, "voom" = dge_voom )
UpSetPlot(DE=DE.list , FoldChange=0 , cutoff =0.01 , type.sig = "FDR")


}
